backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.29.0
python-dotenv==1.0.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
cors==1.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
aiosqlite==0.19.0
PyJWT==2.8.0 